64-bit bug
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/vncbenchtools/trunk@1336 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ rfb.h
@@ -25,7 +25,7 @@
 
 typedef unsigned char  CARD8;
 typedef unsigned short CARD16;
-typedef unsigned long  CARD32;
+typedef unsigned int   CARD32;
 
 #define CONCAT2(a,b) a##b
 #define CONCAT2E(a,b) CONCAT2(a,b)
